Improve journal entry object model for better terminal display
This all started because I didn't like that the FilePath AND the
EntryName were both displayed in the terminal when displaying the index.
It was totally redundant given that journal entry names are expressed
as the current date. I initially attempted to address this issue by using
access modifiers (internal for things I don't want to show in terminal)
but PS doesn't respect public versus internal modifiers. Next I tried to
create a subtype object model, where the lowest ancestor type contained
the data to display. This too was problematic because it limits your ability
to control the order in which properties are displayed in the terminal.
(I think PS will show super type properties first, in the order layed out
in the class, followed by the subtype in the order layed out in the class.)
In the end, I created discrete `Entry` types, with some duplication across
the types. The JournalIndex creation process is now generic, depending
on the level of detail required by the index.