Enable opening journal entries via pipeline
Added a new IJournalEntry parameter to the Open-JournalEntry cmdlet which
enables piping in a journal entry from Get-JournalIndex or Get-JournalIndexByTag.
I had to modify the JournalWriter method that constructs valid entry
paths so that it no longer throws an exception if the entry already
exists. That was a bad design anyway because the method just implies
that it GETS a path, not that it throws if it exists already. I added a
separate, dedicated method for checking whether a entry already exists
on disk. Lastly, I refactored the Open-JournalEntry cmdlet to use the
JournalWriter for constucting paths, rather than having duplicate logic.