Allow AppVeyor to build all branches
I only want master, PRs, and tags to trigger builds, but the last release
I created didn't result in a built because the HEAD commit on master is
also the HEAD commit on another branch which wasn't explicitly white-listed
and apparently that means the build won't run. So this change is an attempt
to fix that situation.