Temp: Fix bug that prevents -Location parameter from being used
This bug was encountered while developing another feature. It was fixed
in that feature branch, but the changes were based on commits with additional
work which would conflict with what's currently deployed to production.
Instead of cherry picking the fix from the other branch, I manually applied
the changes here in order to get it deployed immediately. Afterward, I
will revert this commit and rebase the feature branch on top of the reverted
commit in order to avoid conflicts.